SUBMARINE-268. Unify all test class file name with "*Test" pattern
